Added outline of db structure tests
These tests are basicaly modelled off the test driven graph database
development section in the graph databases book.

They make use of the neo4j libraries directly as that seems to be the
easiest way to create a temporary database, but the actual db code will
probably make use of neocons.