Fixed bug in find-line function.
The find-line functon had a stupid assumption that lines can't move from
their original line number.  This obviously can happen when lines are
added or removed - all subsequent lines have their numbers changed.

Fixing this assumption and instead using the contents of the line
allows db import to successfully complete on the current stubly
repository.

There's still a couple of bugs in find-line that I've documented, but
they don't seem to turn up in a repository of this simplicity.